feature: reorganize and split tests
