+++ .gitignore
@@ -17,3 +17,4 @@
 /tests/fc1.log
 /tests/fc2
 /tests/fc2.log
+**/*.mdb
