+++ Cargo.toml
@@ -114,6 +114,7 @@ farcaster_core = { git = "https://github.com/farcaster-project/farcaster-core", 
 anyhow = "1"
 serde_yaml = { version = "0.8", optional = true }
 toml = { version = "0.5", optional = true }
+env_logger = "0.7"
 
 amplify = "3"
 amplify_derive = "2"
