+++ tests/utils/mod.rs
@@ -1,55 +1,15 @@
 //! Various utils to help build functional tests.
 
-use crate::config::NodeConfig;
+#![allow(dead_code)]
 
+pub mod assert;
 pub mod config;
+pub mod fc;
+pub mod misc;
 
-#[test]
-fn utils_cfg_node_with_proto_serde() {
-    let s = "---\nhost: \"localhost\"\nport: 123\nprotocol: \"http\"\n";
-    let node: NodeConfig = serde_yaml::from_str(&s).unwrap();
-    assert_eq!(
-        node,
-        NodeConfig {
-            host: "localhost".to_string(),
-            port: 123,
-            protocol: Some("http".to_string()),
-        }
-    );
-}
-
-#[test]
-fn utils_cfg_node_with_proto_display() {
-    let node = NodeConfig {
-        host: "localhost".to_string(),
-        port: 123,
-        protocol: Some("http".to_string()),
-    };
-    let display = node.to_string();
-    assert_eq!(display, "http://localhost:123");
-}
-
-#[test]
-fn utils_cfg_node_without_proto_serde() {
-    let s = "---\nhost: \"localhost\"\nport: 123\n";
-    let node: NodeConfig = serde_yaml::from_str(&s).unwrap();
-    assert_eq!(
-        node,
-        NodeConfig {
-            host: "localhost".to_string(),
-            port: 123,
-            protocol: None,
-        }
-    );
-}
-
-#[test]
-fn utils_cfg_node_without_proto_display() {
-    let node = NodeConfig {
-        host: "localhost".to_string(),
-        port: 123,
-        protocol: None,
-    };
-    let display = node.to_string();
-    assert_eq!(display, "localhost:123");
+pub fn setup_logging(level: Option<log::LevelFilter>) {
+    let _ = env_logger::builder()
+        .is_test(true)
+        .filter_level(level.unwrap_or(log::LevelFilter::Info))
+        .try_init();
 }
